implemented triangles::getBaseWidth() and triangles::getHeight() methods, now sf::Vector2f relative position[6] is a variable in the namespace's scope
