+++ src/graphic.cpp
@@ -39,36 +39,29 @@ void createTriangles(const Flock& flock, sf::VertexArray& triangles) {
   assert(static_cast<int>(triangles.getVertexCount()) == flock.getFlockSize() * 3);
 }
 
-void rotateTriangle(const std::shared_ptr<Bird>& bird, sf::VertexArray& triangle, const double delta_theta,
-                    const int i) {
+void rotateTriangle(const std::shared_ptr<Bird>& bird, sf::VertexArray& triangle, const double theta, const int i) {
   const sf::Vertex vertex{bird->get_position()()};  // l'operatore () restituisce la conversione da Point a sf::Vertex
 
-  const double theta = bird->get_velocity().angle();
-
-  const sf::Vector2f relative_position[6] = {sf::Vector2f(0, -height_ / 2),
-                                             sf::Vector2f(-baseWidth_ / 2, height_ / 2),
-                                             sf::Vector2f(baseWidth_ / 2, height_ / 2),
-                                             sf::Vector2f(0, (-height_ * 3 / 2) / 2),
-                                             sf::Vector2f((-baseWidth_ * 3 / 2) / 2, (height_ * 3 / 2) / 2),
-                                             sf::Vector2f((baseWidth_ * 3 / 2) / 2, (height_ * 3 / 2) / 2)};
   const int j = 3 * i;
 
   for (int k = j; k < j + 3; ++k) {
     if (std::dynamic_pointer_cast<Boid>(bird)) {
       triangle[k].position =
-          vertex.position +
-          sf::Vector2f(static_cast<float>(relative_position[k - j].x * std::cos(theta + delta_theta) -
-                                          relative_position[k - j].y * std::sin(theta + delta_theta)),
-                       static_cast<float>(relative_position[k - j].x * std::sin(theta + delta_theta) +
-                                          relative_position[k - j].y * std::cos(theta + delta_theta)));
+          vertex.position + sf::Vector2f(static_cast<float>(relative_position[k - j].x * std::cos(theta) -
+                                                            relative_position[k - j].y * std::sin(theta)),
+                                         static_cast<float>(relative_position[k - j].x * std::sin(theta) +
+                                                            relative_position[k - j].y * std::cos(theta)));
     } else if (std::dynamic_pointer_cast<Predator>(bird)) {
       triangle[k].position =
-          vertex.position +
-          sf::Vector2f(static_cast<float>(relative_position[k - j + 3].x * std::cos(theta + delta_theta) -
-                                          relative_position[k - j + 3].y * std::sin(theta + delta_theta)),
-                       static_cast<float>(relative_position[k - j + 3].x * std::sin(theta + delta_theta) +
-                                          relative_position[k - j + 3].y * std::cos(theta + delta_theta)));
+          vertex.position + sf::Vector2f(static_cast<float>(relative_position[k - j + 3].x * std::cos(theta) -
+                                                            relative_position[k - j + 3].y * std::sin(theta)),
+                                         static_cast<float>(relative_position[k - j + 3].x * std::sin(theta) +
+                                                            relative_position[k - j + 3].y * std::cos(theta)));
     }
   }
 }
+
+float getBaseWidth() { return baseWidth_; }
+float getHeight() { return height_; }
+
 }  // namespace triangles
